+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * ![Enhancement][badge-enhancement] Update CSS source file for JuliaMono, so that all font variations are included (not just `JuliaMono Regular`) and that the latest version (0.039 -> 0.044) of the font would be used. ([#1780][github-1780], [#1784][github-1784])
 * ![Bugfix][badge-bugfix] Fix `strict` mode to properly print errors, not just a warnings. ([#1756][github-1756], [#1776][github-1776])
+* ![Enhancement][badge-enhancement] The table of contents in the generated PDFs have more space between section numbers and titles to avoid them overlapping. ([#1785][github-1785])
+* ![Enhancement][badge-enhancement] The preamble of the LaTeX source of the PDF build can now be customized by the user. ([#1746][github-1746], [#1788][github-1788])
 
 ## Version `v0.27.15`
 
@@ -978,6 +980,7 @@
 [github-1716]: https://github.com/JuliaDocs/Documenter.jl/pull/1716
 [github-1727]: https://github.com/JuliaDocs/Documenter.jl/pull/1727
 [github-1743]: https://github.com/JuliaDocs/Documenter.jl/pull/1743
+[github-1746]: https://github.com/JuliaDocs/Documenter.jl/issues/1746
 [github-1748]: https://github.com/JuliaDocs/Documenter.jl/pull/1748
 [github-1750]: https://github.com/JuliaDocs/Documenter.jl/pull/1750
 [github-1751]: https://github.com/JuliaDocs/Documenter.jl/pull/1751
@@ -996,6 +999,8 @@
 [github-1776]: https://github.com/JuliaDocs/Documenter.jl/pull/1776
 [github-1780]: https://github.com/JuliaDocs/Documenter.jl/issues/1780
 [github-1784]: https://github.com/JuliaDocs/Documenter.jl/pull/1784
+[github-1785]: https://github.com/JuliaDocs/Documenter.jl/pull/1785
+[github-1788]: https://github.com/JuliaDocs/Documenter.jl/pull/1788
 <!-- end of issue link definitions -->
 
 [julia-38079]: https://github.com/JuliaLang/julia/issues/38079